Implements a search results page/search component
* Fixes #754
* Fixes #755

Tries to get the admin to extend the AMO search when possible to share
code and hopefully help us keep the admin search up to date.

This is a first pass and hopefully I haven't duped any tests by mistake.

I think it'd be nicer if the render methods could be more shared, I'd
like to work on that but perhaps as a separate PR.

This one is already quite unwieldy.

r?