feat: Add trailing slashes to URLs (fix #843)
This adds a config option to enable a middleware that adds trailing
slashes to all URLs that don't have one. It's only enabled on AMO
for now as Disco Pane definitely have URLs that are declared without
trailing slash and this does the dumbest thing possible which is
redirect all requests.