Fix getAddonsForUsernames so it will work consistently with multiple usernames
Prior to this fix, if one of the usernames passed didn't have any add-ons loaded, then the
whole function returned null, even if there were some add-ons loaded for some of the
usernames. This situation can arise because when add-ons are fetched for a set of
usernames not all add-ons are fetched and loaded, so we can end up with a username
without any loaded add-ons.