I5150 hiding/404ing language tools link on android (#5244)
fixes #5150

hey @willdurand / @bobsilverberg , I actually haven't got an android to officially test this (yet - I will look into this) but when I use android in the url, the link doesn't appear. 

Please let me know if you think this is sufficient or if this needs to be tested on an actual android. In the meantime, I will look into an emulator/set up for this..