I4744 adding key fixes ssr bug (seen on addon detail in AddonsByAuthorsCard) (#5101)
fixes #4744 

if the theme has a description/summary, the ShowMoreCard would be displayed which uses the Card component. I think b/c there were no keys here it was not acting in an unexpected way.  By adding a varying key value, this looks to be fixed..

Both screenshots are after refreshing the page (SSR bug)

Before (everything looks like it's using AddsonByAuthorsCard except the section classes are wrong):
![Alt text](https://monosnap.com/image/MnbAeBkgOnyWcgaa4Gi2gtXAOlMiIS.png)

After:
![Alt text](https://monosnap.com/image/fPycpAmtBeTvqWoVi3JjL0Sgk3hE2U.png)


*Note*: i wasn't loving this key name but i didn't see a lot of options here. We might need to find a more unique name..