I5703 fixing scaled icon flash (#5787)
fixes #5703 

From what I could tell the url was changed in the src *after* the classname was updated and for a split second you see the icon large b/c it takes on the new classname styles:

![Alt text](https://monosnap.com/image/dcCkZmvEUm78sGjbirOsJTihjGtARi.png)

In the header, the only way that I found that fixes this is to change the html structure or to use a key on the image/wrapping div. This might be another reconciliation issue (?). I decided to go the html approach - changing out a div for a span. let me know what you think..

note: i did a little bit of clean up too : )