I6018 hero setup w/ab test (#6157)
fixes #6018 

This is just a start/proposal for ab hoc. It seems like ab testing will always have /need a cookie, so I set it up a round that. Let me know if that is the wrong assumption.. 
Also, I ran into some issues (like with passing in vals to this) so I welcome early feedback : )



note: this still needs tests