Update test to hide flag button for current user
In addition to signing in the user and ensuring the "flag" button is
hidden for their own replies, the test now signs the user out to ensure
that all flag buttons are shown. This workflow allows unauthenticated
users to see that flagging a review is possible, only after signing in.