Prevent AddonRecommendations from dispatching fetchRecommendations unless the add-on is actually changed (#6808)
Fixes #6805 

This was happening because in `componentDidUpdate` we were checking `oldAddon !== newAddon` and although the contents of the objects are identical, they are not the same object, so that statement was evaluating to `true`. I changed it to check `oldAddon.guid !== newAddon.guid`, which works in this case. I wonder how many other places in our code base we've done this same thing? Perhaps it would be worth opening a code quality issue to audit for that?