Preload the respective number of reviews
To prevent a jarring user experience, this patch will allow a 1:1 ratio
of reviews to loading placeholders.

If the number of reviews to load exceeds 25, only show the first 25 as
per the pagination amount.

If the amount of reviews is undetermined, ie. loading the page directly,
generate a random amount of placeholders within 1 to 25 inclusive.