Disable "classic view" of front-end (#7246)
Fixes #6334 

Note that this removes the "View classic desktop site" link from the footer, and all of the code that supports setting the cookie, but it does not actually "disable" the classic view. That needs to be done on the nginx/server side. I'm not sure if there already exists an issue for that, but I've asked at https://github.com/mozilla/addons-frontend/issues/6334#issuecomment-447926818.