Install the correct version when a specific version is set for installAddon (#7339)
Fixes #7332 

Note that this only addresses the issue where the current version is installed no matter which button is clicked. It does not address the issue of the behaviour of all of the buttons (i.e., they all show install status, they all say "Remove" and they all say "Enable", in different scenarios). That may or may not be addressed via a separate issue.