Use the count of add-ons from the collection add-ons list response in the collection detail page (#7398)
Fixes #4789 

Note that this fixes the count of add-ons in a collection as displayed on the Collection Detail page, but it does not fix the number of add-ons displayed per collection on the Collection Listing page. Please see https://github.com/mozilla/addons-server/issues/8065#issuecomment-432020995 for a discussion/explanation for that.

Note also that if you visit the Collection Listing page and observe an incorrect count, then visit the Collection Detail page for a collection, and then visit the Collection Listing page again, the count on the Listing page will now display the correct count. That is because the correct count is loaded via the Detail page.