I7378-Icon star storybook (#7483)
fixes #7378 

~~So I readded IconStar here so it could be set up in storybook. Although, we will officially turn this on in AMO in #7365 (hopefully).~~ 

I've also added the `Rating` ~~and `RatingsByStar`~~ component~~s here too~~ so you can see the 5 star set up and we'll be able to see the changes with the new PR as these files will be updated in that PR as well.


note: while I do like the props grid, since the props are shown in the Component's variations page, maybe this is enough. Especially since the props (with flow and now with HOCs too it seems) aren't always working :/ For now tho, I just commented this out where it wasn't working properly: on Ratings and the RatingsByStar components. 

also, I would recommend looking into using this plugin again:

https://github.com/evgenykochetkov/react-storybook-addon-props-combinations

It was suggested at 1 point by someone and I think it may be easier to use than manually doing all the use cases/mixes of props as we have to do now.