Add font-display: block on the (preloaded) font subset
In Chrome, if you have a "bad" connection, font-display: auto behaves like
font-display: swap, causing potentially a flash of content but also a layout
shift and that can mess with the LCP because Inter is generally wider than
the fallback font. Since we're preloading the font, the behavior we really
want is font-display: block for the subset.