Stop sending Cache-Control: no-store, send max-age and s-maxage
We don't want browsers to cache our pages at the moment, but sending
Cache-Control: no-store prevents proxies and upstream caches from caching as
well. We override that in nginx at the moment, but the more standards-based
approach is to use max-age and s-maxage instead.

This unlocks the possibility to have a CDN in front of addons-frontend.