Un-inline rating stars SVGs
IconStar is 500 to 700 bytes repeated 5 times for each add-on, on a search page
with 25 add-ons, it adds up! Inlining the svg isn't worth it: it makes the
HTML & DOM larger where we could make an HTTP request from the same connection
thanks to HTTP/2. Inline SVGs also make debugging the DOM in tests more
annoying, as you need to print more to see the rendered results you're
interested in.

Could be further improved by using masks, but for now let's go back to external
SVGs.