Fix add-on rating stars in feedback form
The `Rating` component renders a `div` element, which cannot be placed
inside a `p`, though React is happy to force that to happen on the
client side while browsers "fix" this mistake in the HTML returned by
the server. That explains the short flash of content (with the stars
moving from the middle of the header to the right side in LTR mode).