Update wrong platform copy in add-on details pages when client app is "android"
We already had the right message for non add-on detail pages, so removing
the message specific to the add-on detail page is enough to address the
issue.

This causes us to show what we want, which is, on android add-on detail
pages visited from a Desktop browser, a link to download Firefox for
Android and a link to switch to the desktop version of the page.