Replace smoke test with a basic test against __frontend_lbheartbeat__
We only really need care to check if the docker image express server
works, so we use an endpoint that doesn't require the API (or even
the built statics) with a simpler setup.

Because it's lighter it can be included in the default job that is
ran even on PRs.

`tox` is also removed as it's just not necessary anymore.