Add a default color to dropdown menu items so that the ellipsis shows up (#13612)
.DropdownMenu-items sets a white background, so setting a contrasting
color as well ensures text is shown correctly even when the parent
would set a white color.

This didn't affect links and section headers which sets their own color
but was affecting the '...' ellipsis text that is triggered when the
menu item is too long, which happens in some locales.