Introduce a noStyle prop to the `Card`
This moves the logic to "unstyle" a card in the `Card` component. That
also allows to fix all the cards inside the detail page in a unique way
(e.g. the footer link is now looking nice for all cards).

This patch also removes the almost useless `photonStyle` prop/css.