Restyle the content/rating section of the detail page (#13877)
* Introduce a noStyle prop to the `Card`

This moves the logic to "unstyle" a card in the `Card` component. That
also allows to fix all the cards inside the detail page in a unique way
(e.g. the footer link is now looking nice for all cards).

This patch also removes the almost useless `photonStyle` prop/css.

* Add horizontal separator between add-on description and ratings sections

* Remove external icon on the 'learn more' link in the PermissionsCard

* Add some more space between the description and the ratings when they are side by side