Use 302 instead of 301 when redirecting for an invalid lang
We are enabling/disabling locales depending on whether their
completion percentage so a permanent (301) redirect could be
problematic.

The caching duration could still be an issue, but at least
the non-permanent nature of the redirect should make things
easier, and it's more consistent with addons-server which was
already using a 302 (with that same caching duration).